refactor: extract has-comment-above
