+++ src/rules/no-unsafe-type-assertion.ts
@@ -1,4 +1,5 @@
 import { TSESTree } from "@typescript-eslint/utils";
+import hasCommentAbove from "src/utils/ast/has-comment-above";
 
 import { createEslintRule, type RuleModule } from "../utils/create-rule";
 
@@ -25,7 +26,7 @@ const rule: RuleModule<Options> = createEslintRule<Options, MessageIds>({
   defaultOptions: [],
   create: (context) => {
     function isUnsafeAssertion(node: TSESTree.TSAsExpression): boolean {
-      // Check if it's a double assertion (e.g., as unknown as Type or as any as Type)
+      // Check if it's a assertion through unknown/any (e.g., as unknown as Type or as any as Type)
       if (node.expression.type === TSESTree.AST_NODE_TYPES.TSAsExpression) {
         const innerAssertion = node.expression;
         const innerType = innerAssertion.typeAnnotation;
@@ -39,53 +40,9 @@ const rule: RuleModule<Options> = createEslintRule<Options, MessageIds>({
       return false;
     }
 
-    function hasExplanatoryComment(node: TSESTree.Node): boolean {
-      const { sourceCode } = context;
-
-      // Get the start line of the entire assertion expression
-      let targetNode = node;
-      while (
-        targetNode.parent?.type === TSESTree.AST_NODE_TYPES.TSAsExpression
-      ) {
-        targetNode = targetNode.parent;
-      }
-
-      const comments = sourceCode.getAllComments();
-
-      // Check for comments in the lines immediately before the assertion
-      return comments.some((comment) => {
-        const commentLine = sourceCode.getLocFromIndex(comment.range[0]).line;
-        const commentEndLine = sourceCode.getLocFromIndex(
-          comment.range[1],
-        ).line;
-
-        const nextToken = sourceCode.getTokenAfter(comment, {
-          includeComments: true,
-        });
-
-        if (!nextToken) return false;
-
-        const nextTokenLine = sourceCode.getLocFromIndex(
-          nextToken.range[0],
-        ).line;
-
-        // For multi-line comments, use the end line instead of the start line
-        const effectiveCommentLine =
-          comment.type === TSESTree.AST_TOKEN_TYPES.Block
-            ? commentEndLine
-            : commentLine;
-
-        return (
-          comment.value.trim().length > 0 &&
-          nextTokenLine === effectiveCommentLine + 1 &&
-          nextToken.range[0] <= targetNode.range[0]
-        );
-      });
-    }
-
     return {
       TSAsExpression(node) {
-        if (isUnsafeAssertion(node) && !hasExplanatoryComment(node)) {
+        if (isUnsafeAssertion(node) && !hasCommentAbove(node, context)) {
           context.report({
             node,
             messageId: "noUnsafeTypeAssertion",
